(PUP-3634) Hide password hash from process list for useradd
When managing passwords with useradd provider, the password hash appears
for a quick time when listing running processes (ps, top, strace)

Now the password is set via the chpasswd command that uses stdin to
receive the password from a temporary file. This way it won't appear on
the process list.